server/requirements.txt
defusedxml==0.4.1
Django==1.10.3
django-braces==1.10.0
django-cors-middleware==1.3.1
django-filter==1.0.0
django-oauth-toolkit==0.10.0
django-rest-framework-social-oauth2==1.0.4
djangorestframework==3.5.3
oauthlib==1.0.3
PyJWT==1.4.2
python-social-auth==0.2.21
python3-openid==3.0.10
requests==2.12.1
requests-oauthlib==0.7.0
six==1.10.0
